refactor(renderer): move city plotting into basemap module
